rename: RateLimiterTest 파일 이름 변경
