+++ module-api/src/test/java/org/example/RateLimiterTest.java
@@ -1,23 +1,16 @@
 package org.example;
 
-import org.example.baseresponse.BaseResponse;
-import org.example.dto.request.ReservationRequestDto;
-import org.example.dto.request.ReservationSeatDto;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.*;
 
-import java.util.List;
-import java.util.Objects;
-
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class ApiApplicationTest {
+class RateLimiterTest {
     @LocalServerPort
     private int port;
