+++ module-api/src/test/java/org/example/service/RateLimiterServiceTest.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.service;
 
 import com.google.common.util.concurrent.RateLimiter;
 import org.example.service.movie.FindMovieService;
@@ -11,7 +11,7 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class GuavaRateLimiterTest {
+public class RateLimiterServiceTest {
     @MockBean
     private FindMovieService findMovieService;
